Implemented updates:
There are separate classes to handle bets and cash.
There is now a Deck Class which contains a list of the 52 cards in the
deck.
Cards dealt are now unique.
BigDecimal has been implemented to ensure accuracy of decimal point in
currency.
NumberFormat is being used to display currency.

+++ BlackJack/src/com/blackjack/Player.java
@@ -1,11 +1,13 @@
 package com.blackjack;
 
+import java.math.BigDecimal;
+
 public class Player {
 	private Hand hand;
 	private String name;
-	private double cash;
+	private BigDecimal cash;
 	
-	public Player(String name, double cash){
+	public Player(String name, BigDecimal cash){
 		this.name = name;
 		this.cash = cash;
 	}
